feat: add example for table writes (#1119)
## What changes are proposed in this pull request?

This PR adds a basic example that demonstrates writing to a Delta table
using `kernel-rs`. I consider this as a first draft with several
opportunities for improvement in future iterations:

- Migrating to an official API for table creation (as there currently is
none AFAIK)
- Allowing JSON input instead of generating dummy data (although this
should be orthogonal to what the user aims to learn with the example)
- Automatically create a directory at the target location if it does not
exist yet (I did not want to change the lib's utility function just for
this example so I left it out for now)

This PR also removes deprecated usage of `SyncEngines` in the example
Readmes.

## How was this change tested?
I tested the example locally but did not add any unit tests as it is not
officially part of the library.

---------

Co-authored-by: R. Tyler Croy <rtyler@brokenco.de>
Co-authored-by: Junseo Yoo <59751754+joonyoo181@users.noreply.github.com>
Co-authored-by: OussamaSaoudi <45303303+OussamaSaoudi@users.noreply.github.com>
Co-authored-by: Zach Schuermann <zachary.zvs@gmail.com>