refactor(catalog-managed): push down _last_checkpoint read into LogSegment (#1204)
## What changes are proposed in this pull request?
Push down reading `LastCheckpointHint` into (one of) `LogSegment`
constructor. This is needed so that the LogSegment constructor can
'decide' if it needs to list - in the presence of a `log_tail` in CCv2
we may not need to list at all. Furthermore, even in the absence of
CCv2, this should change is probably still useful in that the
LogSegment::for_snapshot is called after the only place we read the
`LastCheckpointHint` = reasonable to directly embed it. Note that to
keep our existing unit tests i've refactored the `for_snapshot_impl`
with the same signature as the old one.

For reviewers: i've split into two commits, first does material change,
second updates callsites.

## How was this change tested?
refactor: existing