fix: Provide accurate error when evaluating a different type in LiteralExpressionTransform (#1207)
## What changes are proposed in this pull request?

Provide an accurate error.

When an expression has a different type than the underlying schema, it
will raise the appropriate error.

For context, in Iceberg this would bind to the schema. At evaluation,
when the actual scalar value is outside of the bound of an int. it would
rewrite the expression to `AlwaysFalse()` and short circuit. Otherwise,
it will downcast the scalar to an int, and compare the values.

A common example here is when the table had a promotion from int to
long, and some of the older data still has the field encoded as an
integer.

Closes https://github.com/delta-io/delta-kernel-rs/pull/812
Closes https://github.com/delta-io/delta-kernel-rs/issues/813

## How was this change tested?

By adding a new test.

---------

Co-authored-by: sebastian tia <sebastian.tht@gmail.com>
Co-authored-by: Zach Schuermann <zachary.zvs@gmail.com>
Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>