feat: Allow LargeUTF8 -> String and LargeBinary -> Binary in arrow conversion (#1294)
## What changes are proposed in this pull request?

This PR relaxes type-checking rules to allow `LargeUtf8` conversion to
`STRING` and `LargeBinary` conversion to `BINARY`. With this PR Delta
kernel will allow committing arrow record batches that use `LargeUtf8`
or that are built using `LargeStringBuilder` or corresponding types for
building `LargeBinary`.

The main limitation that non-large variations of string and binary types
impose is that they can hold up to 2 GBs of data, which makes it
impossible to do kernel commits if there is a string or binary column
larger than 2 GB. With this PR clients can just pass `RecordBatch` that
uses `LargeUtf8` or `LargeBinary` as column types.

## How was this change tested?
New UT

Co-authored-by: Nick Lanham <nicklan@users.noreply.github.com>