feat!: Introduce `Expression::Variadic` and `Coalesce` expressions (#1198)
## What changes are proposed in this pull request?

This PR introduces `Expression::Variadic` as a new kind of expression
that takes a list of expressions as input and performs an operation on
them.

This change is necessary to support the `Coalesce` expression, which is
also introduced in this PR. We will need a `Coalesce` to support the
read path for row tracking since the row ID (and similarly, the default
commit version) is calculated as `COALESCE(materialized_row_id,
fresh_row_id)`.

### This PR affects the following public APIs

This PR extends the expression framework with `Expression::Variadic` and
thus requires all implementing engines to adopt the new expression.

## How was this change tested?

New UT.

---------

Signed-off-by: Lennart Behme <44500208+lbhm@users.noreply.github.com>
Co-authored-by: Zach Schuermann <zachary.zvs@gmail.com>