chore!: Modify read_actions to not require callers to know details about checkpoints. (#1407)
This PR proposes refactoring log replay to focus on actions that can
occur in commits. It moves the logic for deciding on reading other
checkpoint only metadata into `read_actions`, to isolate other APIs from
understanding the details of what is needed to expand checkpoints. it
also makes the common case for reading actions in the code (identical
checkpoint schema and commit file schema) the common case on
`log_segment` with a new method for the use-case where schemas are
different.