feat!: new Engine StorageHandler head API (#1465)
## What changes are proposed in this pull request?
In CCv2 we need a head API to fetch metadata about commits after they
are published. This is similar to the
[getFileStatus](https://github.com/delta-io/delta/blob/master/kernel/kernel-api/src/main/java/io/delta/kernel/engine/FileSystemClient.java#L96)
API in the java kernel. This PR (1) adds the new `head` method to the
`StorageHandler` trait, and (2) implements it for the default engine


### This PR affects the following public APIs
New `head` API for `StorageHandler`

## How was this change tested?
New UT for default engine impl